some rearranging on the manifest editor
